added db mode tests for convo traversal
